+++ leather/scales/__init__.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python
 
-from leather.scales.annual import Annual
 from leather.scales.base import Scale
 from leather.scales.linear import Linear
-from leather.scales.monthly import Monthly
+from leather.scales.months import Months
 from leather.scales.ordinal import Ordinal
 from leather.scales.temporal import Temporal
+from leather.scales.years import Years